获取费用信息时插入费用共享表

+++ src/main/java/com/lotut/pms/dao/FeeDao.java
@@ -71,5 +71,5 @@ public interface FeeDao {
 
 	List<String> getFeeTypesByPatentType(int patentTypeId);
 	
-	void insertUserFees(List<Map<String, Integer>> userFeeRecords);
+	void insertUserFees(List<Map<String, Long>> userFeeRecords);
 }
